pyproject.toml
[project]
name = "docent"
version = "0.1.0"
description = "A personal control center for grad school workflows."
readme = "README.md"
authors = [
    { name = "John David K. T. Kudadjie", email = "75898705+Kudadjie@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings>=2.14.0",
    "rich>=15.0.0",
    "typer>=0.24.2",
]

[project.scripts]
docent = "docent.cli:app"

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"
